requirements.txt
# Flask Web 框架
Flask>=2.3.0

# MongoDB 驱动
pymongo>=4.5.0

# 日期时间处理
python-dateutil>=2.8.2

# CORS 支持（可选，用于前端开发调试）
flask-cors>=4.0.0

# HTTP 请求
requests>=2.31.0

# HTML 解析
beautifulsoup4>=4.12.0

# Crawl4AI（可选，用于动态页面爬取）
# crawl4ai>=0.2.0
